Move cleanup functions to cleanup module
