+++ buildpacks/sbt/src/main.rs
@@ -5,11 +5,15 @@
 #![allow(clippy::module_name_repetitions)]
 
 mod build_configuration;
+mod cleanup;
 mod detect;
 mod errors;
 mod layers;
 
 use crate::build_configuration::{create_build_config, BuildConfiguration};
+use crate::cleanup::{
+    cleanup_any_existing_native_packager_directories, cleanup_compilation_artifacts,
+};
 use crate::detect::is_sbt_project_directory;
 use crate::errors::ScalaBuildpackError::{
     AlreadyDefinedAsObject, MissingStageTask, SbtBuildIoError, SbtBuildUnexpectedExitCode,
@@ -18,7 +22,6 @@ use crate::errors::{log_user_errors, ScalaBuildpackError};
 use crate::layers::coursier_cache::CoursierCacheLayer;
 use crate::layers::ivy_cache::IvyCacheLayer;
 use crate::layers::sbt::SbtLayer;
-use buildpacks_jvm_shared::{default_on_not_found, list_directory_contents};
 use indoc::formatdoc;
 use libcnb::build::{BuildContext, BuildResult, BuildResultBuilder};
 use libcnb::data::build_plan::BuildPlanBuilder;
@@ -30,10 +33,8 @@ use libcnb::{buildpack_main, Buildpack, Env, Error, Platform};
 use libherokubuildpack::command::CommandExt;
 use libherokubuildpack::error::on_error as on_buildpack_error;
 use libherokubuildpack::log::{log_header, log_info, log_warning};
-use std::ffi::OsStr;
-use std::fs;
 use std::io::{stderr, stdout};
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::{Command, Output};
 
 pub(crate) struct ScalaBuildpack;
@@ -140,63 +141,6 @@ fn create_sbt_layer(
         },
     )?;
     Ok(sbt_layer.env.apply(Scope::Build, env))
-}
-
-// the native package plugin produces binaries in the target/universal/stage directory which is not included
-// in the list of directories to clean up at the end of the build since a Procfile may reference this
-// location to provide the entry point for an application. wiping the directory before the application build
-// kicks off will ensure that no leftover artifacts are being carried around between builds.
-fn cleanup_any_existing_native_packager_directories(app_dir: &Path) {
-    if let Err(error) = default_on_not_found(fs::remove_dir_all(
-        app_dir.join("target").join("universal").join("stage"),
-    )) {
-        log_warning(
-            "Removal of native package directory failed",
-            formatdoc! {"
-                    This error should not affect your built application but it may cause the container image
-                    to be larger than expected.
-
-                    Details: {error:?}
-                "},
-        );
-    }
-}
-
-fn cleanup_compilation_artifacts(app_dir: &Path) -> std::io::Result<()> {
-    let target_dir = app_dir.join("target");
-
-    let target_dir_files = list_directory_contents(&target_dir)?
-        .filter(|path| match path.file_name().and_then(OsStr::to_str) {
-            Some(file_name) => file_name.starts_with("scala-") || file_name == "streams",
-            None => false,
-        })
-        .collect::<Vec<_>>();
-
-    let resolution_cache_files = default_on_not_found(
-        list_directory_contents(target_dir.join("resolution-cache")).map(|directory_contents| {
-            directory_contents
-                .filter(|path| match path.file_name().and_then(OsStr::to_str) {
-                    Some(file_name) => {
-                        !(file_name.ends_with("-compile.xml") || file_name == "reports")
-                    }
-                    None => true,
-                })
-                .collect::<Vec<_>>()
-        }),
-    )?;
-
-    [target_dir_files, resolution_cache_files]
-        .into_iter()
-        .flatten()
-        .map(|path| {
-            if path.is_dir() {
-                fs::remove_dir_all(path)
-            } else {
-                fs::remove_file(path)
-            }
-        })
-        .collect::<Result<Vec<_>, _>>()
-        .map(|_| ())
 }
 
 fn run_sbt_tasks(
